+++ src/egon/data/datasets/electricity_demand_timeseries/hh_buildings.py
@@ -29,8 +29,8 @@ Both tables are created within :func:`map_houseprofiles_to_buildings`.
 * `society.egon_destatis_zensus_apartment_building_population_per_ha`:
     Lists number of apartments, buildings and population for each census cell.
 
-* `boundaries.egon_map_zensus_buildings_filtered`:
-    List of OSM tagged buildings which are considered in the LV-Grid calculation.
+* `boundaries.egon_map_zensus_buildings_residential`:
+    List of OSM tagged buildings which are considered to be residential.
 
 
 **What is the goal?**
@@ -52,11 +52,11 @@ areas not considered in census 2011.
 The assignment of household demand timeseries to buildings takes place at cell
 level. Within each cell a pool of profiles exists, produced by the 'HH Demand"
 module. These profiles are randomly assigned to a filtered list of OSM buildings
-within this cell. The assignment takes place with replacement of every building
-after drew. Therefore multiple profiles can be assigned to one building, making
-it a multi-household building. Hence, every profile is assigned to a building
-but not every building needs to have a profile assigned as not all are
-residential.
+within this cell. Every profile is assigned to a building and every building
+get a profile assigned if there is enough households by the census data. If
+there are more profiles then buildings, all additional profiles are randomly
+assigned. Therefore multiple profiles can be assigned to one building, making
+it a multi-household building.
 
 
 **What are central assumptions during the data processing?**
@@ -90,7 +90,7 @@ Example Query
             Group By cell_id
         ) as t1
     FULL OUTER JOIN(
-        SELECT cell_id, array_agg(array[cast(hh_5types as char),
+        SELECT cell_id, array_agg(array[cast(hh_10types as char),
          hh_type]) as hh_types
         FROM society.egon_destatis_zensus_household_per_ha_refined
         GROUP BY cell_id
@@ -106,6 +106,7 @@ is made in ... the content of this module docstring needs to be moved to
 docs attribute of the respective dataset class.
 """
 from functools import partial
+import random
 
 from geoalchemy2 import Geometry
 from shapely.geometry import Point
@@ -164,7 +165,7 @@ class BuildingPeakLoads(Base):
 
 def match_osm_and_zensus_data(
     egon_hh_profile_in_zensus_cell,
-    egon_map_zensus_buildings_filtered,
+    egon_map_zensus_buildings_residential,
 ):
     """
     Compares OSM buildings and census hh demand profiles.
@@ -181,7 +182,7 @@ def match_osm_and_zensus_data(
     egon_hh_profile_in_zensus_cell: pd.DataFrame
         Table mapping hh demand profiles to census cells
 
-    egon_map_zensus_buildings_filtered: pd.DataFrame
+    egon_map_zensus_buildings_residential: pd.DataFrame
         Table with buildings osm-id and cell_id
 
     Returns
@@ -203,9 +204,9 @@ def match_osm_and_zensus_data(
     )
 
     # count buildings/ids for each cell
-    buildings_per_cell = egon_map_zensus_buildings_filtered.groupby("cell_id")[
-        "id"
-    ].count()
+    buildings_per_cell = egon_map_zensus_buildings_residential.groupby(
+        "cell_id"
+    )["id"].count()
     buildings_per_cell = buildings_per_cell.rename("building_ids")
 
     # add buildings left join to have all the cells with assigned profiles
@@ -381,22 +382,18 @@ def generate_synthetic_buildings(missing_buildings, edge_length):
     missing_buildings_geom["geom"] = buffer
 
     # get
-    buildings_filtered = Table(
-        "osm_buildings", Base.metadata, schema="openstreetmap"
-    )
+    buildings = Table("osm_buildings", Base.metadata, schema="openstreetmap")
     # get table metadata from db by name and schema
-    inspect(engine).reflecttable(buildings_filtered, None)
+    inspect(engine).reflecttable(buildings, None)
 
-    # get max number of building ids
+    # get max number of building ids from non-filtered building table
     with db.session_scope() as session:
-        buildings_filtered = session.execute(
-            func.max(buildings_filtered.c.id)
-        ).scalar()
+        buildings = session.execute(func.max(buildings.c.id)).scalar()
 
     # apply ids following the sequence of openstreetmap.osm_buildings id
     missing_buildings_geom["id"] = range(
-        buildings_filtered + 1,
-        buildings_filtered + len(missing_buildings_geom) + 1,
+        buildings + 1,
+        buildings + len(missing_buildings_geom) + 1,
     )
 
     missing_buildings_geom = missing_buildings_geom.drop(
@@ -409,7 +406,7 @@ def generate_synthetic_buildings(missing_buildings, edge_length):
 
 
 def generate_mapping_table(
-    egon_map_zensus_buildings_filtered_synth,
+    egon_map_zensus_buildings_residential_synth,
     egon_hh_profile_in_zensus_cell,
 ):
     """
@@ -418,13 +415,14 @@ def generate_mapping_table(
     All hh demand profiles are randomly assigned to buildings within the same
     cencus cell.
 
-    * profiles > buildings: buildings have multiple profiles
+    * profiles > buildings: buildings can have multiple profiles but every
+        building gets at least one profile
     * profiles < buildings: not every building gets a profile
 
 
     Parameters
     ----------
-    egon_map_zensus_buildings_filtered_synth: pd.DataFrame
+    egon_map_zensus_buildings_residential_synth: pd.DataFrame
         Table with OSM and synthetic buildings ids per census cell
     egon_hh_profile_in_zensus_cell: pd.DataFrame
         Table mapping hh demand profiles to census cells
@@ -436,9 +434,21 @@ def generate_mapping_table(
 
     """
 
+    def create_pool(buildings, profiles):
+
+        if profiles > buildings:
+            surplus = profiles - buildings
+            surplus = rng.integers(0, buildings, surplus)
+            pool = list(range(buildings)) + list(surplus)
+        else:
+            pool = list(range(buildings))
+        result = random.sample(population=pool, k=profiles)
+
+        return result
+
     # group oms_ids by census cells and aggregate to list
     osm_ids_per_cell = (
-        egon_map_zensus_buildings_filtered_synth[["id", "cell_id"]]
+        egon_map_zensus_buildings_residential_synth[["id", "cell_id"]]
         .groupby("cell_id")
         .agg(list)
     )
@@ -477,10 +487,12 @@ def generate_mapping_table(
     )
 
     # map profiles randomly per cell
+    # if profiles > buildings, every building will get at least one profile
     rng = np.random.default_rng(RANDOM_SEED)
+    random.seed(RANDOM_SEED)
     mapping_profiles_to_buildings = pd.Series(
         [
-            rng.integers(0, buildings, profiles)
+            create_pool(buildings, profiles)
             for buildings, profiles in zip(
                 number_profiles_and_buildings_reduced["building_ids"].values,
                 number_profiles_and_buildings_reduced[
@@ -490,6 +502,7 @@ def generate_mapping_table(
         ],
         index=number_profiles_and_buildings_reduced.index,
     )
+
     # unnest building assignement per cell
     mapping_profiles_to_buildings = (
         mapping_profiles_to_buildings.rename("building")
@@ -564,20 +577,16 @@ def reduce_synthetic_buildings(
     Id's are adapted to continuous number sequence following
     openstreetmap.osm_buildings"""
 
-    buildings_filtered = Table(
-        "osm_buildings", Base.metadata, schema="openstreetmap"
-    )
+    buildings = Table("osm_buildings", Base.metadata, schema="openstreetmap")
     # get table metadata from db by name and schema
-    inspect(engine).reflecttable(buildings_filtered, None)
+    inspect(engine).reflecttable(buildings, None)
 
     # total number of buildings
     with db.session_scope() as session:
-        buildings_filtered = session.execute(
-            func.max(buildings_filtered.c.id)
-        ).scalar()
+        buildings = session.execute(func.max(buildings.c.id)).scalar()
 
     synth_ids_used = mapping_profiles_to_buildings.loc[
-        mapping_profiles_to_buildings["building_id"] > buildings_filtered,
+        mapping_profiles_to_buildings["building_id"] > buildings,
         "building_id",
     ].unique()
 
@@ -589,8 +598,8 @@ def reduce_synthetic_buildings(
     #         zip(
     #             synth_ids_used,
     #             range(
-    #                 buildings_filtered,
-    #                 buildings_filtered
+    #                 buildings,
+    #                 buildings
     #                 + len(synth_ids_used) + 1
     #             )
     #         )
@@ -716,17 +725,17 @@ def map_houseprofiles_to_buildings():
     -----
     """
     #
-    egon_map_zensus_buildings_filtered = Table(
-        "egon_map_zensus_buildings_filtered",
+    egon_map_zensus_buildings_residential = Table(
+        "egon_map_zensus_buildings_residential",
         Base.metadata,
         schema="boundaries",
     )
     # get table metadata from db by name and schema
-    inspect(engine).reflecttable(egon_map_zensus_buildings_filtered, None)
+    inspect(engine).reflecttable(egon_map_zensus_buildings_residential, None)
 
     with db.session_scope() as session:
-        cells_query = session.query(egon_map_zensus_buildings_filtered)
-    egon_map_zensus_buildings_filtered = pd.read_sql(
+        cells_query = session.query(egon_map_zensus_buildings_residential)
+    egon_map_zensus_buildings_residential = pd.read_sql(
         cells_query.statement, cells_query.session.bind, index_col=None
     )
 
@@ -739,7 +748,7 @@ def map_houseprofiles_to_buildings():
     # Match OSM and zensus data to define missing buildings
     missing_buildings = match_osm_and_zensus_data(
         egon_hh_profile_in_zensus_cell,
-        egon_map_zensus_buildings_filtered,
+        egon_map_zensus_buildings_residential,
     )
 
     # randomly generate synthetic buildings in cell without any
@@ -748,9 +757,9 @@ def map_houseprofiles_to_buildings():
     )
 
     # add synthetic buildings to df
-    egon_map_zensus_buildings_filtered_synth = pd.concat(
+    egon_map_zensus_buildings_residential_synth = pd.concat(
         [
-            egon_map_zensus_buildings_filtered,
+            egon_map_zensus_buildings_residential,
             synthetic_buildings[["id", "grid_id", "cell_id"]],
         ],
         ignore_index=True,
@@ -758,7 +767,7 @@ def map_houseprofiles_to_buildings():
 
     # assign profiles to buildings
     mapping_profiles_to_buildings = generate_mapping_table(
-        egon_map_zensus_buildings_filtered_synth,
+        egon_map_zensus_buildings_residential_synth,
         egon_hh_profile_in_zensus_cell,
     )
 
@@ -805,7 +814,7 @@ def map_houseprofiles_to_buildings():
 setup = partial(
     Dataset,
     name="Demand_Building_Assignment",
-    version="0.0.0",
+    version="0.0.2",
     dependencies=[],
     tasks=(map_houseprofiles_to_buildings, get_building_peak_loads),
 )